# api-docs.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/users/signup":{"post":{"operationId":"UserController_singup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddUserDTO"}}}},"responses":{"201":{"description":""}},"tags":["User"]}},"/users/resendOtp":{"post":{"operationId":"UserController_resendOtp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OtpDTO"}}}},"responses":{"201":{"description":""}},"tags":["User"]}},"/users/login":{"post":{"operationId":"UserController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserDTO"}}}},"responses":{"201":{"description":""}},"tags":["User"]}},"/users/confirmEmail":{"post":{"operationId":"UserController_confirmEmail","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/confirmEmailDTO"}}}},"responses":{"201":{"description":""}},"tags":["User"]}},"/users/login-with-google":{"post":{"operationId":"UserController_loginWithGmail","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginWithGoogleDTO"}}}},"responses":{"201":{"description":""}},"tags":["User"]}},"/users/forget-password":{"post":{"operationId":"UserController_forgetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ForgetPasswordDTO"}}}},"responses":{"201":{"description":""}},"tags":["User"]}},"/users/reset-password":{"post":{"operationId":"UserController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDTO"}}}},"responses":{"201":{"description":""}},"tags":["User"]}},"/users/uplaod":{"post":{"operationId":"UserController_uploadFile","parameters":[],"responses":{"201":{"description":""}},"tags":["User"]}},"/users/uplaod-many":{"post":{"operationId":"UserController_uploadFiles","parameters":[],"responses":{"201":{"description":""}},"tags":["User"]}},"/brands/create":{"post":{"operationId":"BrandController_createBrand","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatedBrandDTO"}}}},"responses":{"201":{"description":""}},"tags":["Brand"]}},"/brands/update-image/{id}":{"patch":{"operationId":"BrandController_updateBrandLogo","parameters":[],"responses":{"200":{"description":""}},"tags":["Brand"]}},"/brands/update/{id}":{"patch":{"operationId":"BrandController_updateBrand","parameters":[{"name":"id","required":true,"in":"path","schema":{"$ref":"#/components/schemas/ObjectId"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBrandDTO"}}}},"responses":{"200":{"description":""}},"tags":["Brand"]}},"/brands/freeze-brand/{id}":{"patch":{"operationId":"BrandController_freezeBrand","parameters":[{"name":"id","required":true,"in":"path","schema":{"$ref":"#/components/schemas/ObjectId"}}],"responses":{"200":{"description":""}},"tags":["Brand"]}},"/brands/restore-brand/{id}":{"patch":{"operationId":"BrandController_restoreBrand","parameters":[{"name":"id","required":true,"in":"path","schema":{"$ref":"#/components/schemas/ObjectId"}}],"responses":{"200":{"description":""}},"tags":["Brand"]}},"/brands/{id}":{"delete":{"operationId":"BrandController_deleteBrand","parameters":[{"name":"id","required":true,"in":"path","schema":{"$ref":"#/components/schemas/ObjectId"}}],"responses":{"200":{"description":""}},"tags":["Brand"]}},"/brands/getBrands":{"delete":{"operationId":"BrandController_getBrands","parameters":[],"responses":{"200":{"description":""}},"tags":["Brand"]}},"/categories/create":{"post":{"operationId":"CategoryController_createCategory","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatedCategoryDTO"}}}},"responses":{"201":{"description":""}},"tags":["Category"]}},"/categories/update-image/{id}":{"patch":{"operationId":"CategoryController_updateCategoryLogo","parameters":[],"responses":{"200":{"description":""}},"tags":["Category"]}},"/categories/update/{id}":{"patch":{"operationId":"CategoryController_updateCategory","parameters":[{"name":"id","required":true,"in":"path","schema":{"$ref":"#/components/schemas/ObjectId"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCategoryDTO"}}}},"responses":{"200":{"description":""}},"tags":["Category"]}},"/categories/freeze-Category/{id}":{"patch":{"operationId":"CategoryController_freezeCategory","parameters":[{"name":"id","required":true,"in":"path","schema":{"$ref":"#/components/schemas/ObjectId"}}],"responses":{"200":{"description":""}},"tags":["Category"]}},"/categories/restore-Category/{id}":{"patch":{"operationId":"CategoryController_restoreCategory","parameters":[{"name":"id","required":true,"in":"path","schema":{"$ref":"#/components/schemas/ObjectId"}}],"responses":{"200":{"description":""}},"tags":["Category"]}},"/categories/{id}":{"delete":{"operationId":"CategoryController_deleteCategory","parameters":[{"name":"id","required":true,"in":"path","schema":{"$ref":"#/components/schemas/ObjectId"}}],"responses":{"200":{"description":""}},"tags":["Category"]}},"/categories/getCategories":{"delete":{"operationId":"CategoryController_getCategories","parameters":[],"responses":{"200":{"description":""}},"tags":["Category"]}},"/sub-categories/create":{"post":{"operationId":"SubCategoryController_createCategory","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatedSubCategoryDTO"}}}},"responses":{"201":{"description":""}},"tags":["SubCategory"]}},"/sub-categories/update-image/{id}":{"patch":{"operationId":"SubCategoryController_updateCategoryLogo","parameters":[],"responses":{"200":{"description":""}},"tags":["SubCategory"]}},"/sub-categories/update/{id}":{"patch":{"operationId":"SubCategoryController_updateCategory","parameters":[{"name":"id","required":true,"in":"path","schema":{"$ref":"#/components/schemas/ObjectId"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSubCategoryDTO"}}}},"responses":{"200":{"description":""}},"tags":["SubCategory"]}},"/sub-categories/freeze-SubCategory/{id}":{"patch":{"operationId":"SubCategoryController_freezeCategory","parameters":[{"name":"id","required":true,"in":"path","schema":{"$ref":"#/components/schemas/ObjectId"}}],"responses":{"200":{"description":""}},"tags":["SubCategory"]}},"/sub-categories/restore-SubCategory/{id}":{"patch":{"operationId":"SubCategoryController_restoreCategory","parameters":[{"name":"id","required":true,"in":"path","schema":{"$ref":"#/components/schemas/ObjectId"}}],"responses":{"200":{"description":""}},"tags":["SubCategory"]}},"/sub-categories/{id}":{"delete":{"operationId":"SubCategoryController_deleteCategory","parameters":[{"name":"id","required":true,"in":"path","schema":{"$ref":"#/components/schemas/ObjectId"}}],"responses":{"200":{"description":""}},"tags":["SubCategory"]}},"/sub-categories/getSubCategories":{"delete":{"operationId":"SubCategoryController_getCategories","parameters":[],"responses":{"200":{"description":""}},"tags":["SubCategory"]}},"/products/create":{"post":{"operationId":"ProductController_CreateProduct","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatedProductDTO"}}}},"responses":{"201":{"description":""}},"tags":["Product"]}}},"info":{"title":"My API","description":"API Documentation using OpenAPI","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"AddUserDTO":{"type":"object","properties":{}},"OtpDTO":{"type":"object","properties":{}},"LoginUserDTO":{"type":"object","properties":{}},"confirmEmailDTO":{"type":"object","properties":{}},"LoginWithGoogleDTO":{"type":"object","properties":{}},"ForgetPasswordDTO":{"type":"object","properties":{}},"ResetPasswordDTO":{"type":"object","properties":{}},"CreatedBrandDTO":{"type":"object","properties":{}},"ObjectId":{"type":"object","properties":{}},"UpdateBrandDTO":{"type":"object","properties":{}},"CreatedCategoryDTO":{"type":"object","properties":{}},"UpdateCategoryDTO":{"type":"object","properties":{}},"CreatedSubCategoryDTO":{"type":"object","properties":{}},"UpdateSubCategoryDTO":{"type":"object","properties":{}},"CreatedProductDTO":{"type":"object","properties":{}}}}}